fix: use adapter->tx_buf + sizeof in vmtw, update COAP_RESOURCE_DEFINE, uint64_t unix timestamps for Y2K38
- adapter.c/h: replace stack out[] with adapter->tx_buf in enqueue_next_pending
- coap_*.c: consistent lichen_coap_server per service/linker/ownership
- hal.h, app_interface.h, coap_status.h, puck_location.c + tests: uint32_t -> uint64_t for unix_time/fix_time_unix
- 3 codereview passes per chunk (adapter, each coap, timestamp); all P0-P4 findings filed+addressed as codereview beads
- west build coap + edhoc/oscore/timestamp tests pass
- no dead code, follows patterns

+++ lichen/subsys/lichen/coap/include/lichen/coap_status.h
@@ -73,7 +73,7 @@ struct lichen_coap_dodag_state {
  */
 struct lichen_coap_time_state {
 	bool wall_clock_valid;
-	uint32_t unix_time;       /**< Unix timestamp (valid if wall_clock_valid) */
+	uint64_t unix_time;       /**< Unix timestamp (valid if wall_clock_valid) */
 	const char *source_class; /**< "gnss", "network", "local-client", etc. */
 	const char *source_name;  /**< Human-readable source name */
 	uint32_t age_s;           /**< Seconds since last time sample */
